pyproject.toml
[tool.pytest.ini_options]
asyncio_mode = "auto"
env_files = "deployments/local/env_files/test.env"
filterwarnings = "ignore::DeprecationWarning"

[tool.ruff]
line-length = 99
fix = true
extend-exclude = ["migrations"]

[tool.ruff.lint]
select = ["ALL"]
ignore = ["D", "TRY", "EM"]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S101", "ANN401"]

[project]
name = "api"
version = "0.1.0"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic==1.14.*",
    "asyncpg==0.30.*",
    "fastapi==0.115.*",
    "pydantic[email]==2.10.*",
    "pydantic-settings==2.7.*",
    "pytz>=2025.1",
    "sqlalchemy==2.0.*",
    "uvicorn==0.34.*",
    "greenlet==3.1.*",
    "pyjwt==2.10.*",
    "passlib[bcrypt]==1.7.*",
    "types-pytz==2025.1.*",
]

[dependency-groups]
dev = [
    "faker==36.1.*",
    "pre-commit==4.1.*",
    "pytest==8.3.*",
    "pytest-asyncio==0.25.*",
    "pytest-cov==6.0.*",
    "pytest-dotenv==0.5.*",
    "pytest-mock==3.14.*",
    "testcontainers[postgres]==4.9.*",
]
